refactor: 지하철 노선 command, load 서비스 분리
