+++ src/main/java/subway/web/LineController.java
@@ -3,7 +3,8 @@ package subway.web;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.application.service.LineUseCase;
+import subway.application.service.LineLoadService;
+import subway.application.service.in.LineCommandUseCase;
 import subway.web.request.LineCreateRequest;
 import subway.web.request.LineUpdateRequest;
 import subway.web.response.CreateLineResponse;
@@ -15,28 +16,30 @@ import java.util.stream.Collectors;
 @RestController
 public class LineController {
 
-    private final LineUseCase lineUseCase;
+    private final LineCommandUseCase lineCommandUseCase;
+    private final LineLoadService lineLoadService;
 
-    public LineController(LineUseCase lineUseCase) {
-        this.lineUseCase = lineUseCase;
+    public LineController(LineCommandUseCase lineCommandUseCase, LineLoadService lineLoadService) {
+        this.lineCommandUseCase = lineCommandUseCase;
+        this.lineLoadService = lineLoadService;
     }
 
     @PostMapping("/lines")
     public ResponseEntity<CreateLineResponse> createLine(@RequestBody LineCreateRequest stationRequest) {
-        Long createdLineId = lineUseCase.createLine(stationRequest.toDomain());
-        CreateLineResponse response = CreateLineResponse.from(lineUseCase.loadLine(createdLineId));
+        Long createdLineId = lineCommandUseCase.createLine(stationRequest.toDomain());
+        CreateLineResponse response = CreateLineResponse.from(lineLoadService.loadLine(createdLineId));
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
     @GetMapping("/lines/{lineId}")
     public ResponseEntity<LineResponse> lines(@PathVariable Long lineId) {
-        LineResponse line = LineResponse.from(lineUseCase.loadLine(lineId));
+        LineResponse line = LineResponse.from(lineLoadService.loadLine(lineId));
         return ResponseEntity.status(HttpStatus.OK).body(line);
     }
 
     @GetMapping("/lines")
     public ResponseEntity<List<LineResponse>> lines() {
-        List<LineResponse> lineResponses = lineUseCase.loadLines().stream()
+        List<LineResponse> lineResponses = lineLoadService.loadLines().stream()
             .map(LineResponse::from)
             .collect(Collectors.toList());
 
@@ -45,13 +48,13 @@ public class LineController {
 
     @PutMapping("/lines/{lineId}")
     public ResponseEntity<Void> updateLine(@PathVariable Long lineId, @RequestBody LineUpdateRequest lineUpdateRequest) {
-        lineUseCase.updateLine(lineUpdateRequest.toDomain(lineId));
+        lineCommandUseCase.updateLine(lineUpdateRequest.toDomain(lineId));
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 
     @DeleteMapping("/lines/{lineId}")
     public ResponseEntity<Void> deleteLine(@PathVariable Long lineId) {
-        lineUseCase.deleteLine(lineId);
+        lineCommandUseCase.deleteLine(lineId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
